Fixed "undefined reference to SDL_main"
FINALLY! This problem IS SO HARD TO FIX! I spent several hours trying to test every single solution I have found, and nothing worked. I knew that WMain.cpp is the main entry point, but no matter what combination of #include, #define, int main() definitions I did, nothing is working, I got so frustrated, but I knew I could solve it somehow.

While reading how the build command works, I noticed this funky a** line "Dmain=SDL_main". This line apparently indicates that SDL_main must override the main() method of a program that uses SDL. At first I had absolutely no idea how to remove this line, as nothing in the build indicates this flag, until I found out in my configuration that this flag is included by default, removing it has finally removed the warning! After several hours worth of testing debugging.

But wait, this first solution isn't really applicable for everyone, right? The location of sdl2.pc is in C:\msys64\mingw64\lib\pkgconfig and then removing the -Dmain=SDL_main flag in there, but obviously this global solution isn't very applicable for everyone, as much as possible, this change should be local to this project only, and thankfully there is a way to remove this flag in a localized manner, using filter-out in Makefile.am!

There is also a way to exclude this problematic flag on configure.ac with SDL_CFLAGS=$(shell pkg-config --cflags sdl2 | sed 's/-Dmain=SDL_main//'), but this Makefile.am and Makefile.in workaround should be okay for now.

The reason why ignoring this flag is better than doing something to it (trust me, I tried, NOTHING WORKED) is because the WMain.cpp already handles SDL quite enough with SDL_Init() and SDL_Quit() on the MyApp class, which is what SDL_main is also meant for. Since MyApp already does the job, there is absolutely no need for this method.

That's all! Whew, finally, it is done!