Allow array slider to go up to 16384 elements
Since the new rendering change now allows displaying of large array sizes while the visualization will still look relatively normal, this means that the application should now be able to display up to 16384 elements without having a condensed display.

Yes, loading up even more elements could take a while, but please take note that this is not because of "slow rendering", the step visualization is only rendering elements based on the dynamically calculated step, it never renders anything unnecessary. The reason why larger array sizes take a while to load is because the paint() method needs to scan the entire array, even if it only needs to display a select amount of elements.