Improve inversions calculation for set()
After observing the inversions calculation for a while, I noticed that set() calls RecalcInversions() which counts inversions in O(n^2) time, which is unnecessary. This change makes it so that, first, the inversions of the old value will be counted by checking all numbers in the entire array, when the inversions of the old value are counted, the total number of inversions will now be subtracted with the inversions of the old value. When the old value has been overwritten with the new value, the inversions of the new value will be counted, and then it will be added on the total number of inversions. This should make set() only recount inversions in O(2n) time rather than O(n^2) time, massively reducing overhead.