Fixed inversions calculation
I realized that my initial inversions calculation is incorrect. In order to count the inversions properly, for example, in a sorted array, all numbers at the left side of n are lesser, while all numbers at the right side of n is greater. Using this information, to count the inversions, only either side of n (the number that will be used to "divide" the array) must be evaluated, rather than reading the entire array. 

In this approach, the left side of n will be evaluated to see if any numbers are greater than n, if a number is greater than n and it is on the left side of n, then this number is not on its correct position, because if the array is sorted, all numbers at the left side of n must be lesser than n, therefore, the inverses variable must be incremented by one.