Readded RecalcInversions() call
This is indeed a dilemma, but I did come up with an acceptable conclusion anyway.

The get_mutable() method is an important method for the MyIterator class, as it allows modification of array elements for sorting algorithms that use iterators. However, get_mutable() calls RecalcInversions(), which is an O(n^2) method that counts for inversions of the array. One could say that "well, it just counts for inversions anyway so what's the problem?" but, as always, never underestimate how terrible O(n^2) can be, especially if this code is in a method that is called frequently, or it is a core method.

At first, I thought that removing the RecalcInversions() call will make this method fast, and well yeah it is, since RecalcInversions() must finish first before get_mutable() returns an item. However, removing this call breaks inversions counting for iterator-based sorting algorithms, such as Grail Sort and Pattern-Defeating Quick Sort.

Yes, this is also related to that one other commit I did where I disabled inversions counting for Pattern-Defeating Quick Sort because it performs horribly with it, and this is indeed the reason why, because of the RecalcInversions() call.

So, what now? What can be done? Here is one thing I could say:

It is pretty obvious that the RecalcInversions() call cannot be removed, since it is a "core" function. However, one thing to duly take note is that although manual, the application allows skipping of inversions counting for a specific sorting algorithm, so if one wishes to skip inversions counting for that specific algorithm, then all they have to do is go to SortAlgo.cpp and make the necessary changes. This will avoid the repeated O(n^2) call completely, and this is great if you don't really need inversions counting too. This is especially the case with Pattern-Defeating Quick Sort, in which it performs terribly if inversions counting is allowed for this algorithm.

However, if you do need inversions counting, then you really have no choice but to put up with the repeated O(n^2) call. Removing the RecalcInversions() call completely is not a good idea at all, since you do have the choice of disabling inversions counting for a specific sorting algorithm if you wish.

Well, that's it, at least I have still made set() to be faster, set() doesn't need to do an O(n^2) operation per call.