Reimplemented Weave Merge Sort
The previous implementation of Weave Merge Sort that I added could visually break on depending on the width of the application. This implementation of Weave Merge Sort is the optimized variant, and it works fully in-place, which should never be dependent on the width of the application now. I did not remove the previous implementation in case someone still wants to see it, it works completely fine, it could just "visually" break on smaller array sizes by basically not updating the visualized array when the sorting process is about to finish.