Removed Yield() and TestDestroy() call
Using Yield() to "pause" the thread to allow other threads to run is not recommended, as the "pause" provided by Yield() is reliant on the OS scheduler and this could lead to unpredictable behavior. Instead, using wxMilliSleep() allows for a more consistent and predictable pausing, while also allowing other threads to do their tasks.

Calling TestDestroy() on its own is not going to take any effect. Since the thread is signaled with Wait(), rather than Delete(), when the thread is done executing something, TestDestroy() has no value here, since it would never become true when used as a condition [if (TestDestroy())], even more so if it is called on its own. Unless the thread will be signaled with Delete(), TestDestroy() here is unnecessary.

I also think that if the thread is signaled to terminate using the boolean flag, not only should the DoDelay() method call the Exit() method of the thread, any other code below this should not be executed anymore since the thread has been signaled to stop, hence the added return statement.

The numBrushes change will now be consistent for both step and non-step visualization rendering.