Added Buffer Partition Merge Sort
Also added InsertSort(), an Insertion Sort method that takes low and high bounds so that it can be used on subarrays. Buffer Partition Merge Sort requires the variant of Insertion Sort where the second index variable (j) must be an int so that it is allowed to go below 0, therefore, InsertSort is now implemented this way. InsertionSort() will now call upon InsertSort().