Compressed Scrambled Tail and Head array inputs
The 6 input types (25, 50, 75, Head and Tail) has now been compressed into just 2 array input types named Scrambled Tail and Head, but with a twist! Each reset on Scrambled Tail and Head cycles through 25%, 50%, and 75% Shuffled, so the user still has the "choice" to choose the percentage of shuffled data that they want.