Simplified loop condition on step visualization
By initializing i_step to 0, i == 0 is not needed anymore. But i == last is still needed in order to make sure that the last element is always included in the drawing process. The last variable is now a const, numBrushes is now also a const, as these variables are not meant to change.