Added Ninthers pivot rule selection
To make up for that horrible mistake, I decided to provide a new pivot rule selection known as "Ninthers", it takes 9 samples from the input array, divides the nine samples into 3 groups, get the median from these 3 groups, and then get the median of the medians of the groups.

To make up for this change, Median of Three is now a separate method and it takes a middle argument. As such, Median of Three pivot selection has been adjusted for this change by now computing the middle value first, and then passing the middle value to the method, which the method will return the median.