Optimized pivot lookup for all the new Median of N
The previous implementations of all the new Median of N keeps the pivot elements directly onto the copy vector, and sorts it, and then checks the index of the element in the middle of the copy vector. I realized that this approach is horribly slow particularly for even larger Median of N, such as Median of Nine, which requires 9 additional conditions just to check what index does the middle pivot belong to.

With this new approach, the indices will now be kept instead of the elements themselves, and the Insertion Sort methods used to sort the elements before has now been adjusted to sort the indices based on their values in the original array. With this approach, all Median of N after Median of Three should now have an even more streamlined approach, and it should be way more efficient now, since after the Insertion Sort is done, the value in the middle of copy vector is returned immediately, no more checking. This approach is suitable since in this project, all Quick Sort implementations need the pivot index, not the pivot element.