Fixed volatile ++/-- warning in C++20
This change makes watch() (and m_watch) to not expect a volatile variable anymore. Apparently, even if watch() receives a non-volatile variable, it still works just fine, this can be observed on all sorting algorithms that use watch() to track an index in real-time.

Credits to @Morwenn for guiding me here. The volatile keyword is supposed to be used if the data that a variable contains could change outside of the program, which only happens on low-level, such as handling hardware addresses and registers. To be honest, I really don't think that say, p, a pivot index in Quick Sort, might be changed outside of the program's knowledge for watch() to really need a volatile index. The volatile keyword has a very specific use case now, and if it is going to be used for "multi-threading" purposes, std::atomic must be used instead as this will provide a true guarantee on multi-threaded environments. The volatile keyword only enforces a reread on memory, but it doesn't really protect from deadlocks or data races. Hence, the volatile keyword here is pointless. It doesn't do its job completely well for multi-threading, if that is the purpose of volatile keyword here. If the purpose of the volatile keyword instead is to indicate that the data watch() and m_watch() might change outside of the program instead, then it is also pointless as that should never happen unless the program experiences some form of data corruption or perhaps, the program was hacked or had its values manipulated externally, which should never happen anyway in most cases too.

This change will allow the project to compile without warnings on C++20, if it is desired to truly make watch() and m_watch expect a volatile keyword, then a volatile copy must be used alongside non-volatile copies every time watch() will be used, but this is too much work, extra variables will be declared which is a waste of memory, and it is not a true fix. The true fix is to use std::atomic as already mentioned above.