watch() and m_watch will now hold std::atomic<ssize_t>
After a lot of research and some consultation, I landed into the conclusion that in exchange for volatile, std::atomic<ssize_t> will be used. This will allow watch() to certainly be thread-safe as std::atomic prevents data races, thread access conflicts, and similar issues. This will make watch() "slightly" harder to use since it will now require an std::atomic<ssize_t>, but using std::atomic isn't too hard, and this will also help the program meet the standards of higher C++ compiler versions.

Due to this change, all sorting algorithms that use watch() now has a dedicated std::atomic<ssize_t> to be used for watch(). Some sorting algorithms such as Quick Sort might have to use std::atomic entirely, but since all sorting algorithms here are single-threaded only, it is most likely not needed. What matters is that watch() is now thread-safe, which is what volatile is most likely meant for, but it is not fit for that purpose.