Fixed "cannot terminate thread: The handle is invalid"
So, this fix certainly has some story to it, my god! It took me weeks to fix this issue and it is finally over!

It took me weeks to know what's happening here. The visualizer is reliant upon DoDelay() to end the thread gracefully. Removing wxThread::Exit() or removing the Exit() call on DoDelay() will kill the visualizer.

The reason why the visualizer is producing the "cannot terminate thread" error is because of AbortAlgorithm() calling Wait() on m_thread *after* m_thread has exited using Exit() (which calls wxThread::Exit()), on lower delays this issue isn't so common, but it still has a chance to appear if Wait() is called when m_thread is already invalidated.

This new change now makes it so that Wait() is never called, on the OnRunButton, if m_thread is not nullptr, and it is not alive, simply clearing the pointer will suffice, as waiting for a !IsAlive() thread will produce the "cannot terminate thread" error (which is what also happens in AbortAlgorithm()).

The most major change here is AbortAlgorithm(). Instead of calling Wait(), what AbortAlgorithm() now does is run on a while loop, which checks if m_thread is still alive, if it is alive, the GUI thread will sleep for 1 millisecond, giving DoDelay() time to see the m_thread_terminate flag and exit the thread, once the thread has actually exited with wxThread::Exit(), the thread is no longer alive, making this while loop stop and clearing the pointer.

This fix has been tested on 0.0 ms delay, with array size = 3909, for more than a minute of rapid clicking on the "Random" button using my rapid clicker button, the program doesn't produce the "cannot terminate thread" error. Replacing the while loop with Wait() makes the error come back again.