Fixed "lagging" sound during algorithm reset
This issue isn't very noticeable on low delays, but on high delays, the sound takes a while to disappear after the sorting algorithm has been requested to reset. To resolve this, the sound will now be paused when the Reset button is pressed, and the sound button will be unpressed.

Calling SoundReset() to fix this problem isn't enough, this change will fix it.