Improved DoDelay() algorithm reset responsiveness
The visualizer has another apparent issue where, the higher the delay is, the longer it takes for the algorithm reset to take effect which is not so good, if say the delay is set to 2000 ms, then the application will hang up for 2 seconds before the reset takes effect. This change fixes that by making it so that the reset request now only has a microsecond/millisecond interval, and it will always remain this way regardless of how big the delay is. Even if the delay is set to 1 minute (if TSoS is allowed to have such high delay), the reset will still take effect immediately.