Readded old sleep intervals
This is incase someone wants to revert back to the old sleep intervals, but this will also make the reset requesto take just as long as the configured delay interval. If the delay interval is set to 1 minute, then the application will also hang up for 1 minute to reset the sorting algorithm.