Introduced MilliSleep loop to avoid potential overflow
Although this doesn't completely protect the visualizer from a potential overflow during delay * 1000, this is considered as a partial futureproof solution. 
The way this change works is that if the delay is lower than 1 ms, DoDelay() will switch to microsecond sleeping to accurately represent this delay value, however, if the delay value is bigger than 1ms, DoDelay() will resort to millisecond sleeping to help reduce overhead and avoid potential overflow when delay * 1000 is performed. The millisecond sleep loop will treat values such as 2.65 and 3.25 as whole numbers, but I believe this is a necessary sacrifice as the goal of this change is to avoid calculation overflow.