Drastically increased minimum threshold for Median of N
Credits to @Taihennami on "The Studio" Discord for introducing me to this great minimum threshold formula, as well as introducing me to using Shell Sort for sorting n = 11. The minimum threshold for all higher Median of N is now (n * n), such as 5 * 5 = 25, or n^2, this should allow all higher Median of N pivot rules to perform faster on small array sizes due to having less overhead, as well as providing better pivot selection when the boundary is high enough.