Removed millisecond loop interval in DoDelay()
After some research, I am able to conclude that delay * 1000 isn't going to cause any overflow here in most cases, especially considering that a double can hold an even larger value than an int. But, in case microDelay somehow holds an infinite value, then the additional condition will ensure that microDelay will be set to the max value that a double can hold to ensure that proper comparison and functionality is still there.