Hybrid millisecond/microsecond delay interval
I am very much bothered with just the microsecond loop interval only. 
The reason why it's bothering me so much is that, say, someone allowed 60000 ms in the visualizer, 60000 * 1000 = 60000000, which means the microsecond loop would have to perform a total of 60 million evaluation and incrementation which isn't really good. A previous solution I had was to simply ignore the decimal part and use millisecond interval, so instead of 60 million, only 60000 checks would be performed, but this isn't very good as well since the visualizer fully interprets the decimal part of the ms value.

This struggle finally ends today, finally! The DoDelay() method now has 2 loops, one for the whole number value, and the other one for the fractional value. If the delay for example is set to 1.89 ms, if only the microsecond loop is used, this would result in 1890 iterations in total, with this hybrid approach, only 891 iterations (1 for millisecond, .89 * 1000 = 890 for microsecond) are done which is a good improvement! This slash in iterations becomes more noticeable the larger the ms value is. For example, with 14.89 ms, the resulting iterations with microsecond loop only would be 14890 iterations, with this hybrid approach, it only becomes 904 iterations (.89 * 1000 = 890 + 14) only!

Now I can finally rest easy.